requirements.txt
asgiref==3.2.10
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.2
django-environ==0.4.5
django-heroku==0.3.1
environ==1.0
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.7.5
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
